move mock data to lib
