+++ stories/ProposalsPage/Proposals.stories.tsx
@@ -18,7 +18,7 @@ import { createMockPage } from 'testing/mocks/page';
 import { createMockProposal } from 'testing/mocks/proposal';
 import { builders as _, jsonDoc } from 'testing/prosemirror/builders';
 
-import { members, proposalCategories, spaces, userProfile } from '../../.storybook/lib/mockData';
+import { members, proposalCategories, spaces, userProfile } from '../lib/mockData';
 
 import { ProposalsPageStory } from './ProposalsPageStory';
