+++ tsconfig.json
@@ -122,6 +122,5 @@
   ],
   "include": [
     "**/*.ts",
-    "**/*.tsx"
-  ]
+    "**/*.tsx"  ]
 }
